Rock Scissor Paper
Rock, scissors, paper is a game of hands, between each time two people, where the players simultaneously form one of three objects with their hands, and whoever has the strongest, he wins. The stone is formed by clenching the fist, the scissors with the index and middle fingers extended, and the paper with the palm extended.
The game has two outcomes - either a draw or a win/loss for one of the players.